@@ -1,39 +1,41 @@
 {
-  "name": "@trademe/sliderizr",
-  "version": "0.1.0",
-  "description": "Sliderizr",
-  "main": "index.ts",
-  "dependencies": {
-    "@angular/animations": "^4.4.3",
-    "@angular/common": "^4.4.3",
-    "@angular/compiler": "^4.4.3",
-    "@angular/core": "^4.4.3",
-    "@angular/forms": "^4.4.3",
-    "@angular/http": "^4.4.3",
-    "@angular/platform-browser": "^4.4.3",
-    "@angular/platform-browser-dynamic": "^4.4.3",
-    "@angular/router": "^4.4.3",
-    "core-js": "^2.4.1",
-    "rxjs": "^5.3.0",
-    "zone.js": "^0.8.12"
-  },
-  "devDependencies": {
-    "@angular/cli": "^1.4.3",
-    "@angular/compiler-cli": "^4.4.3",
-    "@types/jasmine": "2.5.38",
-    "@types/node": "~6.0.60",
-    "codelyzer": "~3.1.1",
-    "jasmine-core": "~2.6.2",
-    "jasmine-spec-reporter": "~4.1.0",
-    "karma": "~1.7.0",
-    "karma-chrome-launcher": "~2.1.1",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2",
-    "ts-node": "~3.2.0",
-    "tslint": "~5.3.2",
-    "typescript": "~2.3.3"
-  }
+    "name": "@trademe/sliderizr",
+    "version": "0.2.0",
+    "description": "Sliderizr",
+    "main": "index.ts",
+    "dependencies": {
+        "@angular/animations": "^6.0.3",
+        "@angular/common": "^6.0.3",
+        "@angular/compiler": "^6.0.3",
+        "@angular/core": "^6.0.3",
+        "@angular/forms": "^6.0.3",
+        "@angular/http": "^6.0.3",
+        "@angular/platform-browser": "^6.0.3",
+        "@angular/platform-browser-dynamic": "^6.0.3",
+        "@angular/router": "^6.0.3",
+        "core-js": "^2.5.4",
+        "rxjs": "^6.0.0",
+        "zone.js": "^0.8.26"
+    },
+    "devDependencies": {
+        "@angular/compiler-cli": "^6.0.3",
+        "@angular-devkit/build-angular": "~0.6.8",
+        "typescript": "~2.7.2",
+        "@angular/cli": "~6.0.8",
+        "@angular/language-service": "^6.0.3",
+        "@types/jasmine": "~2.8.6",
+        "@types/jasminewd2": "~2.0.3",
+        "@types/node": "~8.9.4",
+        "codelyzer": "~4.2.1",
+        "jasmine-core": "~2.99.1",
+        "jasmine-spec-reporter": "~4.2.1",
+        "karma": "~1.7.1",
+        "karma-chrome-launcher": "~2.2.0",
+        "karma-coverage-istanbul-reporter": "~2.0.0",
+        "karma-jasmine": "~1.1.1",
+        "karma-jasmine-html-reporter": "^0.2.2",
+        "protractor": "~5.3.0",
+        "ts-node": "~5.0.1",
+        "tslint": "~5.9.1"
+    }
 }
